requirements.txt
elevenlabs
pyaudio
openai
python-dotenv
pillow
langchain-community
requests 